+++ src/driver/amdxdna/amdxdna_gem.c
@@ -7,7 +7,7 @@
 #include <linux/vmalloc.h>
 #include <drm/drm_cache.h>
 
-#include "amdxdna_drv.h"
+#include "amdxdna_drm.h"
 #include "amdxdna_gem.h"
 
 #ifdef AMDXDNA_DEVEL
