--- src/driver/amdxdna/npu3.mk
@@ -1,7 +1,0 @@
-# SPDX-License-Identifier: GPL-2.0
-#
-# Copyright (C) 2024, Advanced Micro Devices, Inc.
-#
-
-# append object files, for example
-#    amdxdna-y += npu3_pci.o
